requirements.txt
Flask==2.3.3
pymongo[srv]==4.5.0
bcrypt==4.0.1
PyJWT==2.8.0
pandas==2.1.1
numpy==1.26.4
openpyxl==3.1.2
requests>=2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7